Revert "muda tipos de variáveis pra coincidir com master"
This reverts commit f78b42eeead421bf6d3dffe37c832660f1f91fbc.

+++ include/wifi.hpp
@@ -9,14 +9,14 @@ namespace Wifi{
 
     typedef struct dataStruct
     {
-        int16_t id;
-        int16_t v;
-        int16_t w;
+        uint8_t id;
+        double v;
+        double w;
     } dataStruct;
 
     void setup(uint8_t robot);
     void OnDataRecv(uint8_t * mac, uint8_t *incomingData, uint8_t len);
-    void receiveData(int16_t *v, int16_t *w);
+    void receiveData(double *v, double *w);
     bool isCommunicationLost();
 
 }
